requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
Pillow==10.1.0
requests==2.31.0
python-decouple==3.8
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
stripe==7.8.0
razorpay==1.4.1
cloudinary==1.36.0
django-storages==1.14.2
boto3==1.34.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.5
django-debug-toolbar==4.2.0
pytest-django==4.7.0
coverage==7.3.2